requirements.txt
flask==3.0.0
flask-cors==4.0.0
rembg==2.0.56
Pillow==10.2.0
gunicorn==21.2.0
onnxruntime==1.17.0